+++ packages/generators/.gitignore
@@ -3,3 +3,4 @@
 !*.test.js
 !/**/*.test.js
 !/templates/*.js
+lib/
