+++ packages/generators/package.json
@@ -2,7 +2,8 @@
   "name": "@webpack-cli/generators",
   "version": "1.0.1-alpha.0",
   "description": "Webpack-CLI generators",
-  "main": "index.js",
+  "main": "lib/index.js",
+  "types": "lib/index.d.ts",
   "keywords": [],
   "author": "",
   "license": "MIT",
@@ -12,6 +13,7 @@
   "dependencies": {
     "@webpack-cli/utils": "^1.0.1-alpha.0",
     "@webpack-cli/webpack-scaffold": "^1.0.1-alpha.0",
+    "@webpack-cli/package-utils": "^1.0.0",
     "chalk": "2.4.2",
     "glob-all": "3.1.0",
     "inquirer-autocomplete-prompt": "1.0.1",
@@ -21,6 +23,7 @@
     "webpack": "4.x.x",
     "webpack-dev-server": "3.8.0",
     "yeoman-generator": "4.5.0"
+
   },
   "peerDependencies": {
     "webpack": "4.x.x"
@@ -42,5 +45,9 @@
   "scripts": {
     "build": "tsc",
     "watch": "npm run build && tsc -w"
-  }
+  },
+  "files": [
+    "lib",
+    "templates"
+  ]
 }
