+++ packages/generators/src/plugin-generator.ts
@@ -1,7 +1,7 @@
 import * as _ from "lodash";
 import * as path from "path";
-
 import addonGenerator from "./addon-generator";
+import * as Generator from 'yeoman-generator';
 
 /**
  * A yeoman generator class for creating a webpack
@@ -11,7 +11,7 @@ import addonGenerator from "./addon-generator";
  * @class PluginGenerator
  * @extends {Generator}
  */
-const PluginGenerator = addonGenerator(
+const PluginGenerator: Generator = addonGenerator(
 	[
 		{
 			default: "my-webpack-plugin",
