+++ packages/generators/src/utils/entry.ts
