+++ packages/generators/src/utils/optionsSchema.json
