+++ packages/generators/src/utils/tooltip.ts
