+++ packages/generators/src/utils/validate.ts
